stylegan2/requirements.txt
scipy==1.3.3
requests==2.22.0
Pillow==6.2.1
h5py==2.9.0
imageio==2.9.0
imageio-ffmpeg==0.4.2
tqdm==4.49.0
tensorflow-gpu==1.14
